+++ android/src/main/java/com/jiguang/jpush/JPushHelper.java
@@ -140,13 +140,11 @@ public class JPushHelper {
             Log.d("JPushPlugin", "the channel is null");
             return;
         }
-
+        Log.d("JPushPlugin", "instance.dartIsReady ="+dartIsReady);
         if (dartIsReady) {
-            Log.d("JPushPlugin", "instance.dartIsReady is true");
             channel.invokeMethod("onOpenNotification", notification);
             openNotificationCache.remove(notification);
         }
-
     }
 
     public  void onNotifyMessageUnShow( NotificationMessage notificationMessage) {
@@ -163,7 +161,10 @@ public class JPushHelper {
     }
     public  void onConnected( boolean isConnected) {
         Log.e(TAG,"[onConnected] :"+isConnected);
-
+        if (channel==null) {
+            Log.d("JPushPlugin", "the channel is null");
+            return;
+        }
         Map<String, Object> results= new HashMap<>();
         results.put("result", isConnected);
        channel.invokeMethod("onConnected", results);