+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## 2.5.0
+更新JCore 440 JPush 520
+## 2.4.9
+修复testCountryCode报错问题
+## 2.4.8
+修复不能清理本地通知问题
+## 2.4.7
+开放requestRequiredPermission API 
+## 2.4.6
+开放setChannelAndSound API 
+## 2.4.5
++ 修复：onConnected 崩溃问题
++ 升级：升级 Android+IOS JPush 5.0.4
+       升级 Android+IOS JCore 4.2.4
++ 新增：setLbsEnable API   
 ## 2.4.4
 + 修复：优化代码结构
 ## 2.4.3